Vector V2 refactor

Created-by: caicongzhi 00516945
Author-id: 118580
MR-id: 15625448
Commit-by: c00516945;z00800224;liujingxiang;y00465900;liqinru
Merged-by: liuyu 00809192
E2E-issues: issue421
Description: + [ ] I've added/changed the documentation
+ [ ] I've added test for the code

V2性能数据
https://onebox.huawei.com/v/b79a6cbdd12ea919ea27feb04c82aedc?type=0

全量构建情况：3条运行失败，其他均成功。
失败用例1：日志输出内容带标签，检查日志中的打印内容，由于打印内容有调整，需要用例跟随适配。功能性实际无问题。
失败用例2：验证Spark Extension的设置内存限制成功，检查日志中的打印内容，由于打印内容有调整，需要用例跟随适配。功能性实际无问题。
失败用例3：Yarn WebUI中Memory Used显示内存申请151G，跟环境相关，多人并发使用存在干扰，重试可成功。
![image](/uploads/b92ed368-7fbf-4e60-b78f-febea2a9a280/1682562201809.png '1682562201809.png')

详细链接
https://integrate-cida.szv.dragon.tools.huawei.com/deskui/project/projectDetail/dc8c3d494a5d480d8f0ef3de4ffe2ce7/33d52c060a8b470cb90fb3b676b96593/treeid?type=group&taskId=6951872514454192128&ticcUrl=http:%2F%2Fhutaf-ticc.inhuawei.com&ticcVersion=V6R13&ticcTenantId=21650e48e4784849a060ea61a843f9db

See merge request: Kunpeng-Computing-DC/BigData/OmniOperatorJIT!485

+++ core/src/codegen/batch_functions/batch_utilfunctions.cpp
@@ -201,7 +201,7 @@ extern "C" DLLEXPORT void CopyDouble(double *dataArray, double *output, int32_t 
     }
 }
 
-extern "C" DLLEXPORT void CopyDecimal128(__int128_t *dataArray, __int128_t *output, int32_t rowCnt)
+extern "C" DLLEXPORT void CopyDecimal128(Decimal128 *dataArray, Decimal128 *output, int32_t rowCnt)
 {
     for (int i = 0; i < rowCnt; i++) {
         dataArray[i] = output[i];
